+++ src/main/java/seedu/duke/commands/SearchCommand.java
@@ -1,8 +1,8 @@
 package seedu.duke.commands;
 
-import seedu.duke.Inventory;
-import seedu.duke.Item;
-import seedu.duke.Ui;
+import seedu.duke.objects.Inventory;
+import seedu.duke.objects.Item;
+import seedu.duke.utils.Ui;
 
 import java.util.ArrayList;
