+++ src/main/java/seedu/duke/objects/Item.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.duke.objects;
 
 import java.util.ArrayList;
