+++ src/main/java/seedu/duke/utils/Storage.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.duke.utils;
 
 public class Storage {
     private String filePath;
