+++ src/main/java/seedu/duke/utils/parser/Parser.java
@@ -1,5 +1,8 @@
-package seedu.duke;
+package seedu.duke.utils.parser;
 
+import seedu.duke.objects.Inventory;
+import seedu.duke.objects.Item;
+import seedu.duke.utils.Ui;
 import seedu.duke.commands.AddCommand;
 import seedu.duke.commands.EditCommand;
 import seedu.duke.commands.FilterCommand;
