+++ src/test/java/seedu/duke/commands/AddCommandTest.java
@@ -1,8 +1,8 @@
 package seedu.duke.commands;
 
 import org.junit.jupiter.api.Test;
-import seedu.duke.Inventory;
-import seedu.duke.Item;
+import seedu.duke.objects.Inventory;
+import seedu.duke.objects.Item;
 
 import static org.junit.jupiter.api.Assertions.*;
