+++ src/test/java/seedu/duke/commands/ListCommandTest.java
@@ -1,9 +1,8 @@
 package seedu.duke.commands;
 
 import org.junit.jupiter.api.Test;
-import seedu.duke.Inventory;
-import seedu.duke.Item;
-import seedu.duke.Ui;
+import seedu.duke.objects.Inventory;
+import seedu.duke.objects.Item;
 
 import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
